% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.8.ball.R
\name{magic.8.ball}
\alias{magic.8.ball}
\title{Magic 8 Ball}
\usage{
magic.8.ball(x)
}
\arguments{
\item{A}{binary response (yes/no) query for the Magic 8 Ball oracle.}
}
\value{
Returns the Magic 8 Ball Oracle's occasionally cryptic response to your query.
}
\description{
Implementation of the Magic 8 Ball invented by Albert Carter & Abe Bookman and originally manufactured by the Alabe Crafts Company.  Later Manufactured by the Mattel toy company.
}
\examples{
magic.8.ball("Will I fulfill my destiny?")
magic.8.ball()
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://en.wikipedia.org/wiki/Magic_8-Ball}
}
\keyword{8}
\keyword{ball}
\keyword{magic}

